Changes requested applied
